group territories when exporting avails csv
